Activar New Degree Fragment
Se activó la creación de un nuevo Degree utilizando Material Design para los estilos. Se añadió verificación de campos. Se actualizó la librería Material a la versión 1.3.0-alpha02. Se modificó el tipo de dato del campo gradeAverage, de la clase Degree,
de float a double. Se creaó recurso XML para manejo de listas desplegables.

+++ app/src/main/java/com/oamorales/myresume/models/Degree.java
@@ -14,11 +14,11 @@ public class Degree extends RealmObject {
     private String discipline;
     private int yearBegin;
     private int yearEnd;
-    private float gradeAverage;
+    private double gradeAverage;
 
     public Degree() {  }
 
-    public Degree(int imageLogo, String degreeTittle, String university, String discipline, int yearBegin, int yearEnd, float gradeAverage) {
+    public Degree(int imageLogo, String degreeTittle, String university, String discipline, int yearBegin, int yearEnd, double gradeAverage) {
         this.id = MyApplication.degreeId.incrementAndGet();
         this.imageLogo = imageLogo;
         this.degreeTittle = degreeTittle;
@@ -81,7 +81,7 @@ public class Degree extends RealmObject {
         this.yearEnd = yearEnd;
     }
 
-    public float getGradeAverage() {
+    public double getGradeAverage() {
         return gradeAverage;
     }
